+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Breaking
 
 - Rename `Persisted::borrow_mut` to `get_mut`
+- Rename `PersistedContainer::get_persisted` to `PersistedContainer::get_to_persisted`
+- Rename `PersistedContainer::set_persisted` to `PersistedContainer::restore_persisted`
 - Persist values on mutation rather than drop for `PersistedLazy`
   - Similar to `Persisted`, `PersistedLazy` now has a `get_mut` method that returns a ref guard